+++ Runtime/EntitySystem/IDamageable.cs
@@ -1,0 +1,14 @@
+using System;
+using CrystalFlux.Core;
+using UnityEngine;
+
+namespace CrystalFlux.EntitySystem
+{
+    public interface IDamageable
+    {
+        void TakeDamage(DamagePacket packet);
+        void TriggerIFrames(float duration);
+        bool IsAlive { get; }
+        event Action<GameObject> OnDeath;
+    }
+}