+++ Runtime/EntitySystem/IKnockbackable.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace CrystalFlux.EntitySystem
+{
+    public interface IKnockbackable
+    {
+        void ApplyKnockback(Vector2 direction, float force, float duration);
+    }
+}