+++ Runtime/EntitySystem/IResourcePool.cs
@@ -1,0 +1,10 @@
+using CrystalFlux.Core;
+
+namespace CrystalFlux.EntitySystem
+{
+    public interface IResourcePool
+    {
+        bool TrySpend(ResourceType type, float amount);
+        bool TryGain(ResourceType type, float amount);
+    }
+}