+++ Runtime/IOnHitEffect.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace CrystalFlux.Core
+{
+    public interface IOnHitEffect
+    {
+        void OnHit(GameObject projectileOwner, GameObject target, Vector3 hitPosition);
+    }
+}