+++ Runtime/ProjectileSystem/ISummonTrigger.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace CrystalFlux.ProjectileSystem
+{
+    public interface ISummonTrigger
+    {
+        bool TrySummon(Vector2 position);
+    }
+}