+++ Runtime/SkillTree/ISkillPointHolder.cs
@@ -1,0 +1,9 @@
+namespace CrystalFlux.SkillTree
+{
+    public interface ISkillPointHolder
+    {
+        int SkillPoints { get; }
+        void AddSkillPoints(int amount);
+        bool TrySpend(int amount);
+    }
+}